requirements.txt
coverage==6.4.4
pytest==7.1.3
pytest-mock==3.8.2
datetime==4.7
Jinja2~=3.1.2
Flask==2.2.2
flask_restful==0.3.9
dicttoxml~=1.7.4
python-simplexml~=0.1.5
flasgger~=0.9.5
peewee~=3.15.3